+++ API/app/src/main/java/com/example/api/FavoriteFragment/FavoritesViewModel.kt
@@ -2,28 +2,29 @@ package com.example.api.FavoriteFragment
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.api.DBandprovider.PersonDb
 import com.example.api.Retrofit.RepositoryAPI
 import com.example.api.Screens
 import com.example.api.repository.RepositorySQLite
 import com.github.terrakok.cicerone.Router
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
 
 class FavoritesViewModel(
-    private val router: Router,
-    private val repositorySQL: RepositorySQLite,
-    private val repositoryAPI: RepositoryAPI
+    private val repositorySQL: RepositorySQLite
 ) : ViewModel() {
+    val listFavorite = MutableStateFlow<List<PersonDb>>(emptyList())
 
-    fun getCharactersFavorite() {
-        viewModelScope.launch(Dispatchers.IO) {
-            repositorySQL.insertFavorite(repositoryAPI.list)
-        }
+
+    init {
+        getCharactersFavorite()
     }
 
-    fun observeAllFavoriteData() = repositorySQL.getAllFavoriteData()
-    fun backFragment() {
-        router.backTo(Screens.getHomeFragment())
+    private fun getCharactersFavorite() {
+        viewModelScope.launch(Dispatchers.IO) {
+            listFavorite.value = repositorySQL.getAllFavoriteData()
+        }
     }
 
     fun delFavoritePerson(uuid: Int) {
