+++ API/app/src/main/java/com/example/api/MainActivity/MainActivity.kt
@@ -38,7 +38,8 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         viewModelMain.create()
         setContentView(R.layout.activity_main)
-        supportActionBar?.setDisplayHomeAsUpEnabled(true)
+        this.supportActionBar?.setDisplayHomeAsUpEnabled(true)
+        viewModelMain.getCharacters(1)
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
