+++ API/app/src/main/res/layout/fragment_home.xml
@@ -10,8 +10,8 @@
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerView"
-        android:layout_width="384dp"
-        android:layout_height="780dp"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
         android:layout_marginStart="11dp"
         android:layout_marginTop="15dp"
         android:layout_marginEnd="11dp"
